+++ src/components/CopyToClipboard/copyText.ts
